Move fabric_event_profiler.hpp out of experimental directory
